feat: add collections and recommendations for the recipes page
